Refactor - move function getFailureRates from FaultEventMenu.tsx ti eventModel.tsx
